backend/requirements.txt
Django==5.1 
gunicorn==22.0.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0 
python-dotenv==1.0.1  
django-environ==0.11.2 
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.29.0
whitenoise==6.11.0
google-auth-oauthlib==1.2.1
google-auth==2.34.0
google-api-python-client==2.150.0